Fixed moving items between PlayerShop and Inventory
Because both storages work on the same persistent ItemStorage, we need
to split it logically. This is done by the new ItemStorageAdapter.

+++ src/GameLogic/InventoryConstants.cs
@@ -142,13 +142,13 @@ namespace MUnique.OpenMU.GameLogic
         /// </summary>
         /// <param name="player">The player.</param>
         /// <returns>The size of the inventory.</returns>
-        public static int GetInventorySize(Player player)
+        public static byte GetInventorySize(Player player)
         {
             var size = EquippableSlotsCount +
                 (InventoryRows * RowSize) +
                 (RowsOfOneExtension * Math.Min(player.SelectedCharacter.InventoryExtensions, MaximumNumberOfExtensions));
 
-            return size;
+            return (byte)size;
         }
     }
 }
